requirements.txt
autopep8==2.0.1
boto3==1.26.69
botocore==1.29.69
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
jmespath==1.0.1
mediafire==0.6.1
ply==3.11
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
requests-toolbelt==0.10.1
s3transfer==0.6.0
six==1.16.0
stone==3.3.1
tomli==2.0.1
urllib3==1.26.14
